# sesi-2/response/response.go
package response

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

type JSONSuccessResult struct {
	Code    int         `json:"code" example:"200"`
	Message string      `json:"message" example:"Success"`
	Data    interface{} `json:"data"`
}

type JSONCreatedResult struct {
	Code    int         `json:"code" example:"201"`
	Message string      `json:"message" example:"Created"`
	Data    interface{} `json:"data"`
}

type JSONBadReqResult struct {
	Code    int         `json:"code" example:"400"`
	Message string      `json:"message" example:"Wrong Parameter"`
	Data    interface{} `json:"data"`
}

type JSONIntServerErrReqResult struct {
	Code    int         `json:"code" example:"500"`
	Message string      `json:"message" example:"Error Database"`
	Data    interface{} `json:"data"`
}

func SuccessResponse(c *gin.Context, data interface{}) {
	c.JSON(http.StatusOK, JSONSuccessResult{
		Code:    http.StatusOK,
		Data:    data,
		Message: "Success",
	})
}

func CreatedResponse(c *gin.Context, data interface{}) {
	c.JSON(http.StatusCreated, JSONCreatedResult{
		Code:    http.StatusCreated,
		Data:    data,
		Message: "Created",
	})
}

func FailResponse(c *gin.Context, respCode int, message string) {
	if respCode == http.StatusInternalServerError {
		c.JSON(respCode, JSONIntServerErrReqResult{
			Code:    respCode,
			Data:    nil,
			Message: message,
		})

		return
	}

	c.JSON(respCode, JSONBadReqResult{
		Code:    respCode,
		Data:    nil,
		Message: message,
	})
}